librocketnpu/tests/rknn_sim/requirements.txt
numpy==1.26.4
rknn-toolkit2
